refactor: credentials step move to credential.ts
